feat: add Google Cloud Firestore Connector
# Add FirestoreIO

## Add FirestoreV1
Set of PTransforms for most Firestore V1 RPC Api Methods
https://cloud.google.com/firestore/docs/reference/rpc/google.firestore.v1

Entry point for accessing Firestore V1 Methods is `FirestoreIO.v1()`.

Currently supported RPC Methods:
* `Read`
  * `BatchGetDocuments`
  * `ListCollectionIds`
  * `ListDocuments`
  * `PartitionQuery`
  * `RunQuery`
* `Write`
  * `BatchWrite`

Transactions are currently not supported for writes due to the multi-leg nature
of begin-then-commit.

### Rpc Quality of Service (QoS)
In an effort to not overwhelm the live Cloud Firestore service, client side
RPC Quality of Service has been built into FirestoreV1. All requests: attempted,
successful and failed are recorded and used to determine the rate at which
requests are sent to Cloud Firestore. In the case of `Write` requests, writes
are enqueued and flushed once a batch size threshold is reached, or a bundle
finishes. A combination of Adaptive Throttling[1], 500/50/5 Ramp Up[2] and
historical batch throughput are used to determine the maximum size of a batch.

Per RPC Method metrics will be surfaced allowing for insight into the values
that feed into the QoS System.

For each RPC Method the following counters are available:
* throttlingMs
* rpcFailures
* rpcSuccesses
* rpcStreamValueReceived

## Add FirestoreSdk

An adapter which provides a managed way for the `com.google.cloud.firestore.Firestore`
SDK to be used despite it not being friendly to running in Beam (not Serializable,
OO Future based API). For full details please refer to the javadocs in
`org.apache.beam.sdk.io.gcp.firestore.FirestoreSdk` and
`org.apache.beam.sdk.io.gcp.firestore.FirestoreSdkFn`

## Test Suites

### Unit Tests

No external dependencies are needed for this suite

A large suite of unit tests have been added to cover most branches and error
scenarios in the various components. Test for input validation and bounds
checking are also included in this suite.

### Integration Tests

A GCP Service account with read & write permissions for Cloud Firestore is
required and the `GOOGLE_APPLICATION_CREDENTIALS` environment variable should be
configured to point to the service account file. Alternatively, the integration
test suite can be ran against the Cloud Firestore emulator by setting
`FIRESTORE_EMULATOR_HOST` to the host-port of the running Cloud Firestore
Emulator, and setting `GOOGLE_CLOUD_PROJECT` to the name of the project that
will be used by the emulator.

An integration tests for each type of RPC along with Firestore SDK operation is
present in `org.apache.beam.sdk.io.gcp.firestore.it.FirestoreIT`. All of these
tests leverage `TestPipeline` and verify the expected Documents/Collections are
all operated on during the test.

All integration tests are integrated with JUnit tests lifecycle and will have
any documents or collections created during the course of a test run cleaned up
after the test finishes running (regardless of success). See the
`org.apache.beam.sdk.io.gcp.firestore.it.FirestoreTestingHelper` JUnit rule for
more details.